Engine containment room atmospherics
Changes the airless and space turfs in the engine containment room to
normal plating so its pressurised.

This will cause the tesla engine to be completely hearable, right now
only half the sounds are hearable because the ball itself is in a
vacuum.

It also creates a much more open space to possibly build a supermatter
or something like that while running the station off off solars.

The singularity engine will ofcourse rip open the flooring and cause the
area to vent which causes a large patch of ice that fades after a few
minutes.
There are some concerns for lag here although a local server handled it
withouth an increase in lag.