Gulp file-ordering, file name styling, add debug script
This moves the order of files to be reliant on the gulp-order plugin,
instead of based off of filename alone; As a result, renamed all of the
current JavaScript back to it's proper name, because the numbers in front
were really bothering me.

The reload.bat file will copy the assets folder to your BYOND cache, for
debugging purposes.

New plugins:
 - gulp-order
  - Allows ordering of file streams via file name. Used for concatenated
    JavaScript.