Add the air currents system. This hook into any atmospheric code (LINDA in our case) via 2 entry points and then generate long distance air currents to help with moving gas around in room / corridors etc.
On the current settings, it use about 1/3 the cpu time that linda use in the same scenario (so it makes atmospherics about 33% time heavier) but is about 4 times faster at moving / stabilizing air around.

By changing parameters it can be made to be much faster and dangerous but i figured people would not necessarily enjoy being flung into space the second they open the air lock to a breached room.