Reclassified events based on github feedback
Blob is now "level 5 biohazard", spacevine is now "level 7 biohazard",
viral infection is now "level 7 viral biohazard" and viral outbreak is
now "level 7 major viral biohazard". This is very close to how it was
originally, the only changes being (A) viral infection moving from 5 to
7, which the other viral event was already at anyway, and (B) viral
outbreak being called 'major viral' instead of just 'viral' so outbreaks
and infections are subtly different. This makes it a little easier to
distinguish blob (now the only 5) from the others, and still preserves
the numbered levels, making the handling of virus events slightly more
consistent in the process. If you want to stick with numbered levels,
I'd like to split them up a bit more (as tigercat suggested), but I
cannot make new sound/AI/outbreak*.ogg files, so this is a compromise
which makes blob a little easier to distinguish, makes viruses more
consistent, and does so without requiring new sound files.