split the aircurrents and airmaster cycle in 4. Instead of once every 20 tick (2 sec) , its doing 1/4 of the work every 5 tick.
This has virtually no impact on the total cpu usage of either proc, but it does reduce the lag felt in game by lot. In a moderatly damaged station atmo, air currents proc use about 0.026sec of cpu. But obviously this can go up higher. By splitting it in 4 instead of having that 0.026 sec to soak every 2 second, its a 0.007sec load to handle every 0.5sec.

The air system is about twice that but the same logic applies. As long as their cpu per call does not go above 1sec on the current code, then this version of the code should eliminate any lag spikes felts by the player from those 2 specific processes.

As a free bonus, it does have for impact that randomly the air seem to update faster even if it does not, because you can see tiles change every 0.5 sec instead of 2sec (but its only 1/4 of the tiles youd normally see every 2 sec)