Makes the surgery computer useful
Adds nanoUI to it, and makes the surgery computer announce the patient's
health and beep when it degrades past certain thresholds. Also lets you
customize those tresholds and mute them individually if you want to.

This is my first PR and I did it more so I could get familiar with byond
and git, please tell if something's wrong