Open Toed Boots
Feat. (feet) toeless jackboots.

Inspired by the toeless jackboots on Polaris.
This works the same as gloves. You take wirecutters, click the footwear,
and bam! Changes the sprites, and opens the toe-- letting Unathi and
Tajara wear the footwear if they weren't allowed already.

If you try to cut open the toes of already toe-cut footwear, it won't
let you.
If you try to cut open the toes of footwear that don't support that
functionality, it does nothing.

Also brings match-shoe lighting up to the same standard as glove and boot
cutting.