Adds Orange and Purple Bandanas, Adds Bandana Colouring, Tweaks Taj/Una Bandana Sprites, Refactors Bandana Adjusting
Also refactors bandana adjustment. I made the change because now it won't revert back to a black bandana after you adjust it twice after colouring it with a crayon in the washing machine.

Tested in the following ways:
Multiple washes, multiple recolourings: Description stayed 'The colours are a bit dodgy' and the name was appropriate for the colour. No sprite issues.
Spammed the adjustment, changing from the head to the face slot at random intervals: Same as above, no sprite issues, description was as intended, name of the item was as intended.

The tweak I made to the East and West directions of the Tajara and Unathi bandana sprites were just to make it fit the muzzle better and make it look like the cloth was pulled taut over the mouth as was the case with the Vulpkanin bandana (and the Vox) to a degree. Slightly less ghetto looking.

Enhanced Bandana Adjusting Functionality

Fixes a bug where adjusting a bandana kept it in the same slot
regardless of whether it was meant to be worn as a hat or a mask.
Fixes a bug where adjusting a bandana didn't change whether or not it
obscured your identity, even when you were wearing it as a hat.