More interactions on help intent
A ton of the stuff, really.
The only balance altering thing is the tail interaction.
When aiming at feet with help intent when you 'hug' a Tajaran, Vulpkanin or Unathi you have a 75% chance to have a 50% chance to make them scream. If you fail the 75% roll you get Weaken(5), allowing the vengeful fur/scalebeasts get revenge. It's toned down in many places to avoid spamming.