Prize Counters and Metastation Fixes
Temporarily adds a prize counter to the station
- Cyberiad: Bar
- Metastation: Dorms / Fitness Room
- Also moved around the arcade machine and some lockers to accomodate

Adds a FULLY UPGRADED prize counter to the Ninja Holding Area/Bar
- This one is not temporary

Fixes a TON of Metastation issues
- Fixes #3514
- Fixes #2983
- Fixes #2977
- Fixes #2976
- Had to shift some structure around with maint, but it should work fine
- Fixes #2968
- Moved the Medbay Surplus Storage to off of a maint tunnel north of
viro (no longer connected to medbay directly), and put the psych office
where that had been
- Fixes #2967
- Fixes #2964
- Wasn't sure if this was referring to the merch store computer or the
PDA cartridge vendor, so I went ahead and added the merch store to
cargo's lobby and the PDA cartridge vendor to tool storage (re-organized
there too)
- Fixes #2959
- Fixes #2952
- Magistrate's machine is in the courtroom, everyone else has their
machine in their respective office
- Fixes #2951
- Just replaced the problem tiles completely. Probably will want to make
it prettier another time
- Fixes #2949
- Account's database is in the meeting room, as the one on Cyberiad
apparently could be reached by anyone with that access
- Scattered a few ATMs around, including in the bar, near the HoP's
desk, Mr. Changs, Arrivals, and some other spots.
- Fixes #2948
- Fixes #2947
- The report mentioned toxins mixing among others, but that one seemed
correct already so I didn't mess with it
- Fixes #2946
- Had to re-pipe a couple spots to compensate for the added adapters
- Fixes # 2942
- Moved some stuff around to accommodate, put the food cart in the
freezer
- Fixes #2939
- Also reinforced the floor like how it is on the Cyberiad
- Fixes #2932
- Well, not exactly. Added a megaphone and EFTPOS to sec, since they
have a secVend and their lockers for the belts, but it was already
agreed that this wasn't really an issue, so this is mostly to close it.

There are still a couple map issues left with Meta, but they are things
that might not be map issues (see #2953), the missing sec podbay (#2926,
need a good spot for that), stuff that was incredibly vague (see #2933),
potentially IC things that are not bugs (#2950), or require sprites
(#2937)