Vox-Fitted Back Stuff, Adds Geneticist Duffel, Fixing Typos and Refactoring Back Icon Generation
Using species_fit and sprite_sheets now works properly for items worn on
the back. Previously it was using the item_state name if it existed in
order to generate the on-mob icon, which was wrong. I restructured it to
work the same way as the rest of the icon generation procs.