Fixes genetics disks (#3657)
The dna/record on the disk was a reference to the one in the buffer, not
a copy, so if you changed the one in the buffer, the one on the dick
changed.
Fixed it for cloning too

BTW, yes, I know I left in an implied src in one of the lines I changed.
This is because whoever created that file is obsessed with src, and
unless I fancied going through the file and stripping out hundreds of
them, taking out a few would actually detract from the readability of
the code