Cocoa Spelling and Relative Pathing Cleanup
Standardized the spelling of "Cocoa", as this was inconsistent
- One spot refers to it as cacao still, but that is because it is in the
legit latin scientific name. Everything else is now cocoa (not coco).

Chocolate now will provide some nutrition (and make you fat if you eat a
lot)
- Cocoa Powder (the arguably main component of chocolate) provided
nutrition, but that nutrition mysteriously vanished when you combine it
with milk and sugar. Now chocolate provides the same nutrition as the
cocoa powder in it

Merged a number of cases where foods had both chocolate and cocoa to now
have just chocolate
- In the case where both were present, the volumes were combined to
maintain food sizes

Renamed cocoa pile to chocolate pile, as this was wrong
- The cocoa pile claimed to be a pure cocoa powder pile, but was
actually just chocolate that was dumped on the floor.

Removed a fair amount of relative pathing, but there's still more to go.