C.M.A. Admin/Self Fix
C.M.A. Admin/Self now works with IPCs regardless of customization (or lack of customization) done while still working with all other species.
Tested by using C.M.A. Admin, C.M.A. Self and the mirrors in the bathrooms as an IPC with a non-morpheus body and head, an IPC with default everything, a human, and a vox. Everything worked just fine afterward.