Fixes 3948
When adjusted, masks that normally cover the mouth will not anymore.
When the mask is pushed back into place, it will cover the mouth again.
You won't be able to run internals from masks that aren't covering your
mouth anymore, either. This reverses when you adjust them back to
normal.