Refactor scissors, fix and finalize prefs, fix missing part definition for xion models, etc.
continued: Monitor changing verb will now only show a list of monitor
styles available for the monitor head brand you are using.

For example, if you chose any of the regular monitor heads (default
morpheus or anything suffixed with mtr.) you will get a list of IPC
standard monitor styles and nothing else.
However, if you chose the hesphiastos alt head (which is also a monitor,
but of a different style) the list will be populated with the monitor
styles suitable for that hesphiastos head, but none others (not even the
ipc standard ones, since those don't fit and hesphiastos alt. isn't in
the models_allowed list)