Plasma security now spawn with their gun in the bag
Fixes #4071

Warden, officers, and pod pilot spawn with the taser in their bag.
Head of Security spawns with egun in the bag.
Brig physician spawns with pen light in their bag.
Also made other species pod pilots spawn with the taser in their bomber
jacket, since it now can hold it, and is more consistent with the rest
of the officers

:cl: pinatacolada
fix: Plasmamen security now spawn with their suit slot items in their
bags
tweak: Pod pilots spawn with taser in the suit slot
/:cl: