Final push, magic belt done
This should be everything finished now. Made sprites for the wand belt and animated them too! Sprites for wand of fireball, polymorph, death, resurrection, teleportation, door creation, and nothing because it's also a wand. Also made a sprite for a single soul stone for the soul stone belt. Didn't feel it needed the same treatment as the bandolier.
The only thing that might not be accepted in this commit is that I commented out the syringe and beaker for the botany belt. They clash with the belt as it is but my argument is that they aren't even used in Hydroponics anymore anyways. At the very most you'll use them for the reagents you grow but chances are you won't put them in your belt anyways.