Hydroponic HUDs
Everything is done. Except it doesn't work. I must be missing something somewhere but otherwise I did everything that appeared needing to do. Based it off of mech diag HUD so that might be the issue.
Code as of now does not let you join the server. Need to revert bit by bit and find out what I did wrong