Fixes for conflicts with other PRs merged in the last two weeks
- Fixes compile issue with energy guns after
https://github.com/ParadiseSS13/Paradise/pull/4515
- Fixes compile issue with targeting spacepods after merge of:
https://github.com/ParadiseSS13/Paradise/commit/0dce120cf6266b229ff0da1a4ff85c03afe4c8b6
- Improves UO71 starting room to be easily re-enterable without hacking
the door again, and to discourage people leaving the first room without
using the ID computer. This is partly to guide people in the right
direction, partly to discourage people using the gateway mission as an
unassailable hideout.
- Modifies queen/empress spit so that other spiders are not harmed by
it. Intended to stop AI queens shooting through their fellow spiders and
killing them in the process.
- Rebalances red spider melee damage.
- Adds a PSI emitter (syndie singularity beacon sprite) orderable via
uplink for 16 TC. Only orderable by traitors with hijack, requires time
and a power source. Once fully powered up, draws a boss-level terror
spider to the station.