Spacepod Fix - Equipment Modules (#4637)
* Spacepod Fix - Equipment Modules

Changes the installed_module variable in equipment to a list, allowing
the rest of the code to read it correctly.  Now when equipment is added
and removed from the pod, it updates properly. Also adds any equipment
the pod starts with to the list.

Fixes the issues in:
https://github.com/ParadiseSS13/Paradise/issues/4599

* Changes List Call

Changes:
equipment_system.installed_modules = new/list()
to
equipment_system.installed_modules = list()

* Properly Placement for List Defined

Removes the definition of the variable as list() from spacepod.dm to
equipment.dm, clearing up a little of size from the total code.

* Removes shot_from

Removes the shot_from lines from the code.