Telsa Can Be Depowered
The tesla energy ball can now be powered down. This adds in an extra
check so that the tesla, when it has no additional energy balls, will
continue to lose power, as well as adding a check to delete tesla when
it runs out of energy. Tesla can no longer be used as an infinite power
source even with the Particle Accelerator off, and will no longer remain
a permanent issue if released.

:cl: Twinmold
Fixes: Adds dissipation to tesla when it has no energy balls, so it can
die out when not fed more energy.
/:cl: