Adjustments to Revenant Abilities (Nerf) (#4693)
Changes some variables in Revenant abilities. Not as extreme as
suggested by some people, but definitely a nerf to Revenants.

This change lowers the amount of confusion from Revenants from 50 to 20
per use of Defile.  This also adds in a maximum confusion of 30 that
Defile can give someone. Finally, this lowers the range of malfunction
from 4 tiles to 2 tiles, as it still uses an EMP function which makes it
an instant kill against IPCs, and greatly incapacitating anyone with
mechanical organs (especially heart).

The numbers for confusion also come out to about equal seconds.

:cl: Twinmold
Tweak: Lowers the range of Malfunction from 4 tiles to 2 tiles, due to
ability to insta-kill IPCs/incapacitate those with mechanical hearts.
Tweak: Lowers amount of confusion given from Revenants from 50 to 20 per
Defile.
Fix: Sets a maximum confused amount from the Revenant's Defile ability
(now 30).
/:cl: