Removes Dissipation
Not really something I agree with. But until other ways of making the engine truly more dangerous are added, dissipation is removed. Decreases the amount of power needed to stop the jumps of electricity to properly reflect the the decrease in actual power levels (a factor of 10, 1000 to 100). Also lowers the division number in shock damage to also reflect this change (400 to 40) so it still does the same amount of damage.