Syndicate Bomb Timer Adjustment
This makes it so the Syndicate Bomb and all subgroups of it reflect a
more accurate time remaining on the bomb. Instead of having a minimum of
60, counting down 1 per 2 seconds, it has a minimum of 120, counting
down 2 per 2 seconds. Adjusts the timer point when it starts beeping
louder to reflect this change, so it matches the current amount.

:cl: Twinmold
Tweak: Makes Syndicate Bombs show a more accurate time remaining until
detonation.
/:cl: