Technical difficulties
Decided to scrap the beekeeping HUD for now. Turned out it's way more complicated that I thought. I'll try it again in the future.
Added the Hydro HUD design for Protolathe. Need to test if it works.
Might have made a mistake in Git so commiting in order to pull from Paradise master to see if anything broke.