Allows admins to save their CC characters (#4717)
* Initial Checkin

- Creates two new admin-only jobs, "Nanotrasen Navy Officer", and
"Special Operations Officer". The former is for IC investigations by
admins of situations on the Cyberiad, and spawns in the admin room. The
latter is for leading/overwatching ERT teams, and spawns at the ERT
office.
- Neither of these jobs are accessible to normal players, either on
join, in preferences, or as a random job.
- When they are spawned, no announcement is made, and they do not show
up on the crew manifest.
- In other news, the "select equipment" debug verb has also been
adjusted so that NT Navy Officers get the same equipment either way.

* Fixes teleporter on admin station so that admins can use it to teleport directly to bridge

* Fixes CC jobs not always displaying CC HUD icon for sechuds

* Adds an implant that dusts someone on their death

* Tweaks

Changes:
- CC officers now spawn with an internals box in their backpacks, like
everyone else.
- CC officers now get dust and death alarm implanters in their backpacks
on spawn, rather than automatically having the implants in their bodies.
This enables admins to test things on admin station without worrying
about dusting or having their deaths broadcast.
- Dust implants no longer provide death alarm functionality.
- Dust implants no longer process().
- Dust implants are now based on explosive implants, and can be manually
activated via deathgasp.
- CC officers now spawn with CC headsets that are named 'centcom
headset'.

* Deletes CC encryption key

* Switches to using imp_in.dust()