Stops vehicles unintentionally being space vehicles
So, it turns out you can be pulling a vehicle whilst riding it, and then
when you reach space the vehicle thinks it is being pulled through space
rather than ridden.

I would guess this is why it's been so hard to replicate... People have
pulled the vehicle, then got on it and not realised they're still
pulling it.

Ideally we'd be able to stop people pulling it when riding it, but that
would either mean snowflaking into the pull code (yuk) or refactoring
the pull code, which doesn't seem worth it for just this, so instead
I've just added a check to make sure if it's being pulled through space
it isn't the buckled mob doing it

fixes #4498