Adds warning when client BYOND version is old
Adds a chat text warning when a client connects to the server with a
version of BYOND which is at least one major version behind what is
current.
E.g we've declared the standard to be 510, and they connect with 509.

This warning does not actually stop them doing anything. It just tells
them to go update their BYOND.

This does not support minor versions (e.g: 510.1347) because the
client's "byond_version" var only reports the major version number.
Still, should help remind people who return to the game after an absence
to update their client before playing.