Paradise Pop
Adds in a new Bottler machine used to combine a set of three ingredients
into a beverage!

Adds a new line of drinks, branded Paradise Pop, to be made with the
bottler. Currently, the possible flavors are:
- Paradise Punch
- Apple-pocalypse
- Berry Banned
- Blackeye Brew
- Grape Granade
- Meteor Malt
These beverages currently have no mechanical effect, though that will be
added at a later date along with additional recipes

The bottler can dispense empty glass bottles, plastic bottles, or metal
cans as well as fill these containers with the result of a processing of
ingredients. Successful processing requires both careful selection of
ingredients AND proper ordering, or else the entire drink is rendered a
horrible and nigh-undrinkable disaster.

Used cans and bottles, both from the bottler and the vending machines,
can be reinserted for recycling even when crushed, so pick up that can!
Clown stole all your cans and bottles for his latest prank? Simply slide
in some sheets of glass, plastic, or metal to restock the machine's
reserves!

Order your Bottler today from cargo for only 35 supply points, and we'll
even throw in a free wrench to help you move it to its new home! (will
map one into the bar later)

Minor other changes:
Sugarcane and grapes received kitchen_tag values, so they may be used in
the bottler
- Also allows grapes to be properly juiced again in a blender
Water bottles from the vendors now are plastic bottles, instead of glass