Airlock/Shuttle Compatibility Patch + Cycling Airlocks
This patch makes it so airlock code is now compatible with current
shuttle code. This changes how airlocks work slightly. This also adds in
cycling airlocks to all exterior airlocks other than the prisoner
transfer shuttle airlocks (due to a current PR that would cause
conflicts between this and it if implemented).

This also adds in one additional air vent to each of the two engine
airlocks that connect to the engine room. This also locks the open
trader airlock.