Fixes NTSpecOps Radios
- Changes NT Navy Officer / Special Ops Officer headsets to broadcast to
ERT radio by default, BUT also have the ability to talk to special ops
(deathsquad) radio by using .- (or #-, or #-).
- This fixes a bug where Special Ops Officers can't actually talk to
special ops teams because special ops teams are on a different frequency
to ERTs, and neither of the two channels (ERT or SpecOps) can be
selected with a :X/.X/#X style letter.