The Sleeping Carp Vampires
There is a small chance someone who is a vampire will also get traitor,
or vice versa. If they get the martial art scroll (sleeping carp), they
are unable to harm intent the head to drink their blood anymore. This
makes it so they can if they have sleeping carp.