Fixes Ore Storage for Spacepods
Fixes being able to pick up and unload ore with the ore storage system
for spacepods, since someone forgot to make storage a list. forceMove()
was removed from this code, since it runtimes very badly (because it
doesn't work with lists, apparently).

:cl: Twinmold
Fix: Spacepod ore storage system can now pick up ore and unload it like
it was supposed to.
/:cl: