Pod Disabler and Grab Fixes
Fixes the bug that lets you destroy pods with stamina damage weapons by
adding in a check, as well as fixes the bug that makes it possible to
pull someone out of a pod they are not in anymore.

Fixes #5088

:cl: Twinmold
Fix: Can no longer destroy pods with disablers and other stamina damage
weapons.
Fix: Can no longer pull someone out of a pod when no one is actually in
the pod.
/:cl: