Adds a new axe for Atmos-traitors
Give Atmos-traitors a unique 10TC axe to buy.

The axe, on its first hit, will do an additional 70 burn damage on top
of its normal 30 brute. It will spark and make a rather obvious noise
when it connects. After this one charge is used it will function as a
completely normal axe.

http://i.imgur.com/NKnwXga.gif

Sprites courtesy of Fullofskittles
:cl: IcyV
add: Adds a unique axe for Atmos-traitors
/:cl: