Vox Raider Check Kills for Inviolate
Effectively, adds the check that makes it so you aren't allowed to kill
(many, at least) under the Inviolate. Doesn't check for deaths in the
server itself, only kills in the raiders minds, which the original code
failed to do by adding all deaths to voxkills. As we currently have,
it's a auto-complete objective, which is just a poor fix to the original
issue.

```var/vox_allowed_kills = 3``` gives them a little leeway, but can
easily be changed to any acceptable number.

:cl: Twinmold
Tweak: Vox Objective 4 now calculates the kills vox raiders actually
have, instead of always complete.
/:cl: