Gives Atmos-Traitors a new fireaxe.
Give Atmos-traitors a unique 10TC axe to buy.

This axe starts with a single charge and can gain more charges over time
by destroying windows and grilles. Each grille or window destroyed has a
4% chance of giving another charge to the axe. On a charged hit the axe
will hit for 30 brute and 30 burn. It will otherwise function as a
normal fireaxe if out of charges.

![nknwxga](https://cloud.githubusercontent.com/assets/6348574/17047510/9ab3c986-4f93-11e6-83b8-c3e4598704eb.gif)

Sprites courtesy of Fullofskittles
:cl: IcyV
add: Adds a unique axe for Atmos-traitors
/:cl: