Adds an exploding chameleon flag for traitors
This gives traitors a 7TC chameleon flag. It can be used once to be
disguised as any other flag on the station. It comes with an explosive
payload equivalent to a minibomb. It explodes once the flag finishes
burning. By default it looks like a standard NT flag that has a separate
name and description by default.

Would have re-opened #5281 instead but things went horribly wrong on my
local side.

:cl: IcyV
add: Adds an exploding chameleon flag for traitors.
/:cl: