Fixes Accidental Greentext for Maroon and Redtext for Assassination Objectives.
If you had the maroon objective as a traitor and your target was in any
sort of locker on the shuttle when it docked, it would give you green
text.

Maroon objectives now check if the target's in a locker and, if so,
checks the locker's location in order to correctly evaluate objective
completion.

If you had the assassination objective as a traitor and you killed them,
yet they were on a Z-level greater than the Derelict, it would give you
red text.

Assassinations are now location insensitive.