WOLOLO
New traitor chaplain items: Missionary Robe and Missionary Staff.

They look like ordinary chaplain hoodies, but the Missionary Robes
actually are armored and can be linked to a single Missionary Staff to
recharge "faith" while wearing the robes.

The Missionary Staff looks like either the red or blue godstaff
nullrods, and is also a nullrod (can't be reskinned). However, the true
power comes when linked to a set of Missionary Robes, charging the staff
with "faith" that can be used to perform ranged conversions on
individual targets within sight of the missionary.

Conversions mimic the effects of a mindslave implant, except that it
only last for 10 minutes before the victim returns to normal. The victim
is specifically told when they deconvert that they forget their actions
and master. Conversion is able to overcome loyalty implants sometimes,
but if successful it will take longer than normal to recharge. Likewise,
the more "educated" crew members (psychiatrist and librarian) are more
resistant to conversion, though much less so than the loyalty implanted
crew (these resistances do not stack). Civilians, on the other hand,
have a chance to only use half the "faith" of a normal conversion,
meaning you can attempt another conversion much sooner. Crew that is
already mindslaved (whether by a staff or implant) cannot be converted,
as they are already mentally dominated, so this can also potentially
help you make a new traitor friend if you play your cards right.

Successfully converting a crew member mindslaves them to you for 10
minutes, and recolors their worn jumpsuit to match the color of your
staff (red or blue, chosen randomly when the staff is spawned). Failure
to convert a target expends 25-75 "faith" (based on circumstances),
while a successful conversion typically uses 100 "faith" but can use
more or less in special cases. As long as you wear the Missionary Robes
and keep your Missionary Staff close by, the "faith" recharges passively
and will notify you when you are able to attempt another conversion.

Times, chances, and values are subject to change.

TODO: Upon conversion, play "WOLOLO" to the missionary and new zealot,
to inform them both of the good news. Also, testing.