WOLOLO Cleanup v2
Moves the timer (and deconversion) to the mind datum instead of on the
staff.

Rewrites nullrod reskin list-building to instead rely on a variable set
on the subtype of nullrod to determine whether or not to include it by
default.
- By default, all nullrod subtypes EXCEPT fluff subtypes and the
missionary_staff subtype have this set to TRUE.
- Fluff subtypes still re-add themselves to the list as appropriate
through their fluff_transformations list entries, like they previously
did.
- This also allows future coders to potentially make new nullrod
variants that are not reskin options (such as for admin-only versions,
or special away mission reward versions) with better stats / effects by
simply setting "reskin_selectable = FALSE" in the subtype declaration
instead of manually removing the path from the list