Minor mask tweaks, adds alt_head mask refitting, adds fitted versions of unathi masks for the new alt head.
The alt_head mask refitting system works like this: If the player has an alt_head that's not "None", the system will look for an icon state in the most relevant mask icon file that has the mask's icon state and alt_head suffix ("[wear_mask.icon_state]_[alternate_head.suffix]") and, if it exists ([(alternate_head && ("[wear_mask.icon_state]_[alternate_head.suffix]" in mask_icon.IconStates())) ? "_[alternate_head.suffix]" : ""]") the mask sprite that will appear on the character will have that icon state. If it doesn't exist, the mask sprite that will appear on the character will have the default icon state for the mask object.