Fun Preservation Circuits
Adds a cooldown to the manual reboot command, to prevent anti-fun
comdoms from endlessly rebooting to lock crew out of the game.

In the event of the game server being knocked offline, the manual reboot
can be used even if it was on cooldown from a previous restart. This
still initiates a fresh cooldown.