Additional logic for disposal
disposalholder checks if it has items with mixed tags and if it merges
with a different tag so it redirects back to cargo to be sorted
correctly.
Renamed is_fat_person to tooBig since it's a better name to consider
something too big to process in disposals.
Untagged mail returns back to cargo.
Adds the option to change which tag to default to for delievery chutes
via defaultTag (defaults to 1).
Adds two varribles to disposalholder: mixed for if different destination
tags exist and mail if wrapped packages exist.
Readds the wrapsortjunction to cargo to look out for mixed items and
mail to send to the outlet south of the normal open.
Edited the delivery chute in disposals to send stacked items back to
cargo, fixing issue #5732
Added additional glass to the disposal outlet to prevent items from
going to the other belt, missing the recycler.