Disposal Tweaks & Fixes
Adds defaultTag to Delivery Chutes so a different destination could be
given to items thrown into them.
disposal.dm got a minor tweak and removed a bunch of implied `src`
Updated the map to add reinforced glass to prevent items from flying
onto the other conveyor and updated the chute to direct glass and metal
back to cargo as intended using defaultDestinationTag.