Prevents flicker() via Boo! when shadowling is near.
Proposed change for #5686

Allows Ghosts to cast boo but prevents the flickering of lights when a
Shadowling is near.  The Ghost is _not_ informed that their flicker
hasn't worked because they may not have Antag HUD enabled.

N.B. range parameter on line 32 of range() may need adjusting as I have
approximated it's effectiveness.