Fix pAI Atmosphere Sensor
The problem was that in the old code, data["pressure"] was stored as a
String to round it. While NanoUI was programmed to use
helper.smoothRound() on that again which didn't like the String Object
pressure became.
As all data is integers and rouned in the UI, the fix chosen is to make
data["pressure"] a number again.