Fixes Fake Blood Sample on Hotel Cleaver
The blood sample of the cleaver in the cult room in the space hotel has
an unlisted blood sample, causing a runtime when scanned by the
detective's scanner, breaking said scanner. This fixes the variable by
turning it into a proper list without changing the underlying value. As
such, it does give 1e+009 as the readout, but otherwise works
appropriately with no runtime.

:cl:Twinmold
Fix: Fixes a runtime issue caused by using the detective's scanner on
the space hotel cleaver.
/:cl: