Fixes Wire to Firing Range APC
Simple fix... Would have fixed it with the other PR had I been aware of
the issue. If there are other mapping issues as well, I can include them
to this PR just so we have fewer PRs filling up the space for mapping
fixes.

:cl:Twinmold
Fix: Wiring to brig firing range now a node instead of a curved wire.
/:cl: