Pod Fixes and Some Runtime Fixes
Fixes sending pod pilots to the void when the pod explodes/is destroyed,
as well as giving them damage warnings, just like passengers get. Also
fixes it so if a guardian traps something, then dies, doesn't runtime.
Lastly, fixes it so there isn't a randomly runtime from opening a
critter crate, as well as cleaning up a little snowflake code, thanks to
Crazylemon.