Fixes some genetics issues.
Markings will now load correctly from UI and skin tone changes will now
be immediately shown.

Fixes a bug where changing identities to a Vulpkanin with all the
markings as a changeling wouldn't load the markings, and another where
it'd take a body icon update to show the change of skin tone if you were
to change identity to a human.

This is also a likely fix to a bug where, upon changing identities, a
changeling's head would be invisible... Spooky!