Removes PDA Chatrooms
Removes @tastyfish's PDA chatrooms.

While a nice feature, it's mainly only used as a tool for greytiders to
grief the telecomms server by adding everybody to a chatroom. THis
causes the comms to duplicate and forces command to shut down the PDA
server until it clears up.

It's also rarely ever actually used by anybody in a legitimate sense.

Debate down below if you want, a fair few people are behind me with
this.