Fixes an Issue with Cloned Vampires Losing Powers
Alright, so the root of the issue was how the vampire gamemode tracks
the owner of vampire stuff. This can be seen in line 197 of vampire.dm
that it tracks a body much like the mind datum does. However, before my
fix, this 'owner' fell out of date when a vampire got cloned. While the
vampire would get cloned, this owner variable would still refer to their
old body, and so stuff would break.

Now, it gets updated with the current body when the vampire gets cloned,
thus resolving the issue.