PR to get us up to date with paradise #2 (#13)
* Less Blurry Character Preferences images. Ported from VG

https://github.com/d3athrow/vgstation13/pull/11149

* Patreon Modifications

 - Adds a `donators` table for our patreon supporters.
   - Intended to be half-automated, half-manual, in order to correlate
     patreon accounts to ckeys. Uses the `active` field so that supporters
     don't have to have their ckey associated every time they stop and
     restart their donations.
 - Current actions:
   - Donators at tier 1 or greater get a palm tree next to their name in OOC
     (toggleable)
   - Donators at tier 1 or greater get +5 loadout points.
     - Known issue: Loadout points will not update if the table is
       modified during a round where the user has already logged in, due
       to preference datum caching. There's not really any good way to fix
       this, unfortunately.

* OOPS FORGOT THE DEBUG MESSAGES WEW

* ERP Warning Fix

* Add donator loadout category and item defines

* Sharper character creation preview images, tails/body accessories now show up in ID photos.

* Pod Fixes and Some Runtime Fixes

Fixes sending pod pilots to the void when the pod explodes/is destroyed,
as well as giving them damage warnings, just like passengers get. Also
fixes it so if a guardian traps something, then dies, doesn't runtime.
Lastly, fixes it so there isn't a randomly runtime from opening a
critter crate, as well as cleaning up a little snowflake code, thanks to
Crazylemon.

* Added FreeStylaLTs items, OOC tag rounding

Added FreeStylaLT's donor items, changed OOC tag to make it more round,
fixed FreeStylaLT's donor items by adding them to paradise.dme

* Changes Sound and Message Dup Codes to Procs

By request of Crazylemon. Doing this also removed the need to have the
check for passengers and pilots from the health checks of the pod, as
they are no longer necessary and unified inside the new procs.

* Err... Should Still be !health

Apparently it's just so that single sound plays when the pod is out of
health utterly.

* Makes the admin logout tracking system actually work

* Adds 2 other spots to notify the main chat

* Add cost = 2 to fur cape and fur coat

* Fixes some genetics issues.

Markings will now load correctly from UI and skin tone changes will now
be immediately shown.

Fixes a bug where changing identities to a Vulpkanin with all the
markings as a changeling wouldn't load the markings, and another where
it'd take a body icon update to show the change of skin tone if you were
to change identity to a human.

This is also a likely fix to a bug where, upon changing identities, a
changeling's head would be invisible... Spooky!

* Automatic changelog generation for PR #5926 [ci skip]

* Automatic changelog compile, [ci skip]

* Cryo chambers no longer let you use massive beakers of SF

at a cost of only 1u per application - instead, `reaction_mob` will only
be called with the amount of reagents that actually entered the person's
body

* Timer Controller Update

* Adds SHODAN Fluff

What it says on the title really.

For captain role only. Styla has confirmed.

* Automatic changelog generation for PR #5928 [ci skip]

* Automatic changelog generation for PR #5915 [ci skip]

* Automatic changelog generation for PR #5914 [ci skip]

* Automatic changelog compile, [ci skip]

* Automatic changelog generation for PR #5895 [ci skip]

* Automatic changelog compile, [ci skip]

* Revert "Makes the admin logout tracking system actually work"

* Revert "Revert "Makes the admin logout tracking system actually work""

* Makes NukeOps Pinpointers wearable in PDA slots

* Fixes an Issue with Cloned Vampires Losing Powers

Alright, so the root of the issue was how the vampire gamemode tracks
the owner of vampire stuff. This can be seen in line 197 of vampire.dm
that it tracks a body much like the mind datum does. However, before my
fix, this 'owner' fell out of date when a vampire got cloned. While the
vampire would get cloned, this owner variable would still refer to their
old body, and so stuff would break.

Now, it gets updated with the current body when the vampire gets cloned,
thus resolving the issue.

* Fixes a bug with Changeling Regenerative Stasis

Changelings who initiated the stasis while they were alive and were
'killed' during the waiting period for regenerative readiness will now
regenerate properly.

* Fix AI static exploit

* Changes the direct setting of stat to a more sane one called enacted through a proc.

* Automatic changelog generation for PR #5944 [ci skip]

* Automatic changelog compile, [ci skip]

* Automatic changelog generation for PR #5942 [ci skip]

* Automatic changelog compile, [ci skip]